requirements.txt

torch
Pillow
basicsr
realesrgan
numpy
